Added a test to assess the quality of the approximate math functions.

+++ src/approx_math.h
@@ -24,12 +24,16 @@
 /**
  * @brief Approximate version of expf(x) using a 4th order Taylor expansion
  *
- * The absolute error is of order 10^-6 for -0.2 < x < 0.2.
+ * The absolute error is smaller than 3 * 10^-6 for -0.2 < x < 0.2.
+ * The absolute error is smaller than 2 * 10^-7 for -0.1 < x < 0.1.
+
+ * The relative error is smaller than 1 * 10^-6 for -0.2 < x < 0.2.
+ * The relative error is smaller than 4 * 10^-8 for -0.1 < x < 0.1.
  *
  * @param x The number to take the exponential of.
  */
 __attribute__((always_inline)) INLINE static float approx_expf(float x) {
-  return 1.f + x * (1.f + x * (0.5f + x * (1.f / 6.0f + 1.f / 24.0f * x)));
+  return 1.f + x * (1.f + x * (0.5f + x * (((float)(1.0 / 6.0)) + ((float)(1.0 / 24.0)) * x)));
 }
 
 #endif /* SWIFT_APPROX_MATH_H */
